FPGA: Add `new_unbooted` test case (#314)
This is a basic test case that boots the FPGA and checks the MCI boot
status.

This change also removes "fpga_realtime" as an enabled dependency in the xtask crate.

Since the dependency features of a workspace are resolved via a union,
this enabled "fpga_realtime" for the rest of the workspace.